feat(mcp): ship the MCP server in the installed binary as `xerj mcp`
XERJ had a complete, tested MCP server that no user could obtain. The
crate exists (engine/crates/xerj-mcp, 10 tools, 25 unit tests), it is a
workspace member, and CI builds and smoke-tests it — but release.yml
builds only `-p xerj-server`, so no release artifact ever contained it,
`landing/get` installs only `xerj`, and no agent-facing document
mentioned it. The only way to run XERJ's MCP server was to clone the
repository and compile it. A 22-run first-contact agent study found this
was the single biggest structural gap: shell-less agent personas
(MCP-only desktop assistants, function-calling hosts) failed 100% of
runs, and "no native MCP" is a recurring field complaint.

Mechanism: a `xerj mcp` subcommand on the one binary, not a second
release artifact. An agent can install exactly one thing; `xerj
autoindex` and `xerj brain` already establish the subcommand pattern;
and, decisively, this needs no change to the release matrix, to
`landing/get`, or to `get.ps1`. Everyone who has already run
`curl -fsSL https://xerj.org/get | sh` gains an MCP server on their next
upgrade with nothing else to download. A second artifact would have
added eight cross-compiles and an installer branch for a binary the
installer does not put on PATH.

Refactor: src/main.rs → src/lib.rs with a `pub async fn run(&[String])`
entry point, plus a thin src/main.rs that calls it, so the standalone
`xerj-mcp` binary keeps working exactly as CI drives it today. The
dispatch sits in `main()` ahead of the production runtime, next to the
`__extract-pdf` worker, for two reasons: a stdio proxy needs two worker
threads rather than 8×cores at 4 MiB of stack each, and MCP reserves
stdout exclusively for the JSON-RPC stream — everything below that point
(`parse_args`'s --help fast path, the banner, the console link) writes
to stdout, nothing above it does.

Cost to the shipped binary: zero new third-party dependencies. xerj-mcp
needs tokio, reqwest, serde_json and anyhow; reqwest is already a
non-dev dependency of xerj-autoindex and xerj-ai, both of which
xerj-server already pulls in. `run` also gained `--url` / `--auth`
flags (the environment stays the documented path) and now rejects
unknown arguments instead of ignoring them — a typo'd flag that is
silently dropped leaves the agent querying the wrong node with no way
to notice.

The published schema was stale and nothing compared it: the file
advertised six tools while the binary served ten, missing every
xerj_brain_* tool. landing/docs/agents/schemas/mcp-tools.json is now
GENERATED from a live `tools/list` response by
scripts/mcp-schema-check.sh (--write), and gated twice: that same script
in check mode, wired into ci.yml next to the MCP smoke step and driving
`xerj mcp` so the gate covers the path users actually have; and
engine/crates/xerj-mcp/tests/published_schema_drift.rs, which pins the
file to `tool_specs()` field-by-field and asserts the subcommand
dispatch still exists. Restoring the six-tool file makes the script exit
1 with a diff — verified by hand.

Verified end-to-end against a live node (private ports 9421/9422/9423,
three fixture docs plus a semantic_text + dense_vector index):
initialize → tools/list returned all ten tools, and all ten executed
against real data — xerj_search (1 hit, score 0.949), memory_store →
memory_recall (round trip, score 1.726), brain_link → brain_overview →
brain_ego → brain_unlink (edge 183e4394…, live 1 → invalidated 1),
semantic_search over semantic_text (2 hits), vector_search (cosine 1.0
on the exact vector), hybrid_search (rrf). Engine errors pass through
verbatim: a semantic query against a plain `text` field came back as the
engine's own 400, unmodified, as isError:true.

Docs: README.md (setup, tool table, copy-pasteable mcpServers block),
AGENTS.md (agent-facing section plus a Where-to-look row),
landing/docs/cli.html (a `xerj mcp` reference section; its opening line
"There are no subcommands" was already false before this change and now
lists all four). Every config block uses an ABSOLUTE binary path,
because an MCP host launched from a desktop icon does not inherit a
shell PATH and a bare "xerj" fails there with no useful error.

Honesty preserved: xerj_semantic_search is described as server-side
embedding with the LEXICAL feature-hashing embedder unless the node runs
--embed-mode neural, and the xerj_brain_* tools keep both mandated
honesty sentences (verified present in the generated schema).

Files: engine/crates/xerj-mcp/{Cargo.toml,src/lib.rs,src/main.rs,
tests/published_schema_drift.rs}, engine/crates/xerj-server/{Cargo.toml,
src/main.rs}, scripts/mcp-schema-check.sh, .github/workflows/ci.yml,
landing/docs/agents/schemas/mcp-tools.json, landing/docs/cli.html,
README.md, AGENTS.md.
