+++ engine/crates/xerj-server/Cargo.toml
@@ -24,6 +24,9 @@ xerj-compress = { path = "../xerj-compress" }
 xerj-cluster  = { path = "../xerj-cluster" }
 xerj-console-api     = { path = "../xerj-console-api" }
 xerj-autoindex = { path = "../xerj-autoindex" }
+# `xerj mcp` — the MCP stdio server, in the one binary users install. Pulls in
+# only reqwest + serde_json, both already in this tree.
+xerj-mcp     = { path = "../xerj-mcp" }
 
 tokio.workspace = true
 axum.workspace = true
